django/requirements.txt
Django==3.0.4
djangorestframework==3.11.0
psycopg2
django-cors-headers==3.2.1
djangorestframework-jwt==1.11.0
pykafka==2.8.0
kafka-python==2.0.1